# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "nipa crm api",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "NIPA server"
        }
    ],
    "paths": {
...
        "/v1/keywords": {
            "get": {
                "tags": [
                    "Keywords"
                ],
                "summary": "show all keywords",
                "operationId": "674f923992c0bf9df9edf6324900da50",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Keywords"
                ],
                "summary": "create a new Keyword",
                "operationId": "0b6c3a5160c535c3ac231615da77ef43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Keyword"
                                ],
                                "properties": {
                                    "keyword": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "hello"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/v1/keywords/{id}": {
            "get": {
                "tags": [
                    "Keywords"
                ],
                "summary": "get a Keyword by id",
                "operationId": "7c2cb70ecc0581521e8d0225942e59c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Keyword",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Keywords"
                ],
                "summary": "update a Keyword by id",
                "operationId": "9ca769f7ded02c71e1124cc003441d58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Keyword",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Keyword"
                                ],
                                "properties": {
                                    "keyword": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "english"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Keywords"
                ],
                "summary": "delete a Keyword by id",
                "operationId": "67b0a8cf21717f68083137bd97d96e69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Keyword",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}